+++ packages/core/src/index.ts
@@ -14,4 +14,5 @@ export * from "./sourceReferences.js";
 export * from "./sourceVisuals.js";
 export * from "./templates.js";
 export * from "./designPacks.js";
+export * from "./director.js";
 export * from "./typography.js";